smtp-server: import tests from the previous implementation
